Create screenpy.py
This is a simple Python wrapper that builds more functionality to the Bash screen command. The command is less than friendly and while this wrapper doesn't entirely replace it, it does make it much easier to rejoin a screen that is active ... 

It's worth pointing out that this wrapper only handles one part of the screen command process - and I may expand upon this and simply create one more Python element that launches a fresh terminal instance using the command this is based on - but that isn't really necessary. For now, this should help anyone who is frustrated with the inefficient process of rejoining a session running over screen.