Update screenpy.py
I have made substantial changes with this script - and I think the end user will find these changes helpful. There as an issue with earlier versions of this where the program would invoke an error message from a lack of output validation - and this had to do with a failure to check whether or not a screen session was even running to begin with. I fixed that problem.